requirements.txt
Django>=4.2.0,<5.0
channels>=4.0.0
gunicorn>=21.2.0
django-allauth>=0.54.0,<0.55.0
Pillow>=10.0.0
whitenoise>=6.5.0 